+++ src/Optional.Abstractions/NetStandardCompatibility/IsExternalInit.cs
@@ -1,7 +1,9 @@
 ﻿using System.Diagnostics.CodeAnalysis;
 
 // ReSharper disable once CheckNamespace
+#pragma warning disable IDE0130 // Namespace does not match folder structure
 namespace System.Runtime.CompilerServices;
+#pragma warning restore IDE0130 // Namespace does not match folder structure
 
 // ReSharper disable once UnusedType.Global
 /// <summary>